api/requirement.txt
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.8.0
matplotlib
numpy